requirements.txt
Flask==2.3.2
supabase==1.0.4
python-dotenv==1.0.0
requests==2.31.0
flask-cors==4.0.0
bcrypt==4.0.1
flask-session==0.5.0
